Add support for display-name configuration property
There was some confusion over the semantics of stable-name. This was not
helped by the documentation in picker-input.sample.scm.

This patch teases apart the property of a node where the "name-of-the-node"
is guaranteed to be stable across graph versions and the property where the
"names" of the node are displayable to the picker user.

Originally these two properties were mostly synonymous because I hadn't
thought of a good way to provide stability for the node-names of the
non-register data. The non-register data were the synonyms and spelling
mistakes and we didn't want to show spelling mistakes or politically
sensitive synonyms to the user.

Now these two properties are properly independent. We add another colour to
the Graphviz rendering such that it shows nodes that are displayable but
not canonical in Grey.

There remains a small amount of ambiguity around the semantics of canonical
nodes that are not also displayable.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>