Add DWP synonyms but don't consume them
These are only the synonyms that aren't already part of the existing
"Combined Synonyms".

This list possibly needs working on. There appears to be some outright
mistakes (eg: "`Cote dIvoire`"), some typographical differences
(eg: "`O’zbekstan`"), some capitalization issues (eg: "`République
française`") and some examples of what appears to be multiple synonyms in
one field (eg: "`中国/中华`").

We don't know enough to comment on the correctness of this. For the time
being we're adding it "as-is".