Make it easier to work out what new col-spec entries are needed
When new columns are added to the spreadsheet we get an error when we try
to handle them.

Here we enhance the error message so that we can work out the whole list of
new columns and update col-spec in one go. We go to some lengths to format
the column names nicely.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>